+++ .idea/misc.xml
@@ -3,6 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/Fabiola/AndroidStudioProjects/BottomsUp/app/src/main/res/drawable/ic_home_black_24dp.xml" value="0.15625" />
         <entry key="..\:/Users/Fabiola/AndroidStudioProjects/BottomsUp/app/src/main/res/layout/activity_main.xml" value="0.14114583333333333" />
         <entry key="..\:/Users/Fabiola/AndroidStudioProjects/BottomsUp/app/src/main/res/layout/fragment_home.xml" value="0.1" />
         <entry key="..\:/Users/Fabiola/AndroidStudioProjects/BottomsUp/app/src/main/res/layout/view_holder_category.xml" value="0.1" />
