+++ app/src/main/res/layout/fragment_browse.xml
@@ -6,18 +6,15 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="16dp"
-    tools:context=".ui.home.HomeFragment">
+    tools:context=".ui.browse.BrowseFragment">
 
     <TextView
-        android:id="@+id/text_home"
+        android:id="@+id/text_browse"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:textAlignment="center"
+        android:layout_margin="8dp"
         android:textSize="20sp"
-        tools:text="Featured"
+        tools:text="Browse by category..."
         app:layout_constraintTop_toTopOf="parent"
         />
 
@@ -26,7 +23,7 @@
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:scrollbars="vertical"
-        app:layout_constraintTop_toBottomOf="@id/text_home"
+        app:layout_constraintTop_toBottomOf="@id/text_browse"
         app:layout_constraintBottom_toBottomOf="parent"
         />
 </androidx.constraintlayout.widget.ConstraintLayout>
