+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -2,9 +2,9 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/navigation_home"
+        android:id="@+id/navigation_browse"
         android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
+        android:title="@string/title_browse" />
 
     <item
         android:id="@+id/navigation_dashboard"
