+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -3,13 +3,13 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_home">
+    app:startDestination="@+id/navigation_browse">
 
     <fragment
-        android:id="@+id/navigation_home"
-        android:name="com.example.bottomsup.ui.home.HomeFragment"
-        android:label="@string/title_home"
-        tools:layout="@layout/fragment_home" />
+        android:id="@+id/navigation_browse"
+        android:name="com.example.bottomsup.ui.browse.BrowseFragment"
+        android:label="@string/title_browse"
+        tools:layout="@layout/fragment_browse" />
 
     <fragment
         android:id="@+id/navigation_dashboard"
