+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,8 @@
 <resources>
     <string name="app_name">Bottoms Up</string>
-    <string name="title_home">Home</string>
+    <string name="title_browse">Browse</string>
     <string name="title_dashboard">Dashboard</string>
     <string name="title_notifications">Notifications</string>
+
+    <string name="browse_by_category">Browse by category…</string>
 </resources>
